requirements.txt
streamlit==1.32.0
pandas==2.1.0
numpy==1.26.0
scikit-learn==1.4.0
shap==0.44.1
evidently==0.4.22
anthropic==0.25.0
plotly==5.18.0
matplotlib==3.8.0
seaborn==0.13.0
fpdf2==2.7.9
python-dotenv==1.0.0
scipy==1.12.0
xgboost==2.0.3
imbalanced-learn==0.12.0
